Fixes for the types of various iterators as used in driver programs

+++ src/driver/neurotrees_copy.cc
@@ -303,14 +303,14 @@ int main(int argc, char** argv)
   {
     // Create packed representation of the tree subset arranged per rank
     vector<char> sendbuf; 
-    vector<int> sendcounts, sdispls;
+    vector<size_t> sendcounts, sdispls;
     data::serialize_rank_tree_map (size, rank, tree_rank_map,
                                    sendcounts, sendbuf, sdispls);
     tree_rank_map.clear();
     
     // Send packed representation of the tree subset to the respective ranks
     vector<char> recvbuf; 
-    vector<int> recvcounts, rdispls;
+    vector<size_t> recvcounts, rdispls;
     throw_assert(mpi::alltoallv_vector(all_comm, MPI_CHAR, sendcounts, sdispls, sendbuf,
                                        recvcounts, rdispls, recvbuf) >= 0,
                  "neurotrees_copy: error while sending tree subset to assigned ranks");
